pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "google-genai>=1.19.0",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "statsmodels>=0.14.4",
    "uvicorn>=0.34.3",
]
